fixed some php notices

+++ camagru-dev/app/models/Home.php
@@ -57,6 +57,9 @@ class Home{
     {
         $this->db->query("SELECT `id` FROM `posts` WHERE `img` = :img");
         $this->db->placeholder(":img", $img);
+        $ret = $this->db->single();
+        if (!$ret)
+            return;
         $postId = $this->db->single()->id;
 
 
@@ -81,7 +84,10 @@ class Home{
     {
         $this->db->query("SELECT `id` FROM `posts` WHERE `img` = :img");
         $this->db->placeholder(":img", $img);
-        $postId = $this->db->single()->id;
+        $ret = $this->db->single();
+        if(!$ret)
+            return;
+        $postId = $ret->id;
 
 
 
@@ -104,7 +110,10 @@ class Home{
     {
         $this->db->query("SELECT `id` FROM `posts` WHERE `img` = :img");
         $this->db->placeholder(":img", $img);
-        $id = $this->db->single()->id;
+        $ret = $this->db->single();
+        if(!$ret)
+            return;
+        $id = $ret->id;
 
         $this->db->query("INSERT INTO `comments` (`post_id`, `user_id`, `comment`) VALUES (:id, :userid, :comment )");
         $this->db->placeholder(":id", $id);
@@ -119,33 +128,42 @@ class Home{
     }
 
 
-    public function deleteComment($cmt_id)
+    public function deleteComment($cmt_id, $bimg)
     {
         $this->db->query("SELECT `user_id`, `post_id` FROM `comments` WHERE `id_c` = :cmtid");
         $this->db->placeholder(":cmtid", $cmt_id);
         $ret = $this->db->single();
 
-        $cmt_owner = $ret->user_id;
-        $post_id = $ret->post_id;
-
-        
-        $this->db->query("SELECT `user_id` FROM `posts` WHERE `id` = :post");
-        $this->db->placeholder(":post", $post_id);
-        $post_owner = $this->db->single()->user_id;
-        
-        // echo $cmt_owner."   ".$post_owner;
-        if ($cmt_owner == $_SESSION['id'] || $post_owner == $_SESSION['id'])
+        if($ret)
         {
-            $this->db->query("DELETE FROM `comments` WHERE `id_c` = :cmtid");
-            $this->db->placeholder(":cmtid", $cmt_id);
-            $ret = $this->db->execute();
-            $this->db->query("SELECT `id_c`, `id`,`username`, `comment` FROM `comments` JOIN `users` ON `comments`.`user_id` = `users`.`id` WHERE `post_id` = ".$post_id." ORDER BY `created_at` DESC LIMIT 4");
-            $comments = $this->db->result();
-            $data = new stdClass();
-            $data->c = $comments;
-            $data->pid = $post_id;
-            return $data;
+            $cmt_owner = $ret->user_id;
+            $post_id = $ret->post_id;
+
+            
+            $this->db->query("SELECT `user_id` FROM `posts` WHERE `id` = :post");
+            $this->db->placeholder(":post", $post_id);
+            $post_owner = $this->db->single()->user_id;
+            
+            // echo $cmt_owner."   ".$post_owner;
+            if ($cmt_owner == $_SESSION['id'] || $post_owner == $_SESSION['id'])
+            {
+                $this->db->query("DELETE FROM `comments` WHERE `id_c` = :cmtid");
+                $this->db->placeholder(":cmtid", $cmt_id);
+                $ret = $this->db->execute();
+            }
         }
+        $this->db->query("SELECT `id` FROM `posts` WHERE `img` = :bimg");
+        $bimg = "/img/posts/".$bimg.".png";
+        $this->db->placeholder(":bimg", $bimg);
+        $post_id = $this->db->single()->id;
+
+
+        $this->db->query("SELECT `id_c`, `id`,`username`, `comment` FROM `comments` JOIN `users` ON `comments`.`user_id` = `users`.`id` WHERE `post_id` = ".$post_id." ORDER BY `created_at` DESC LIMIT 4");
+        $comments = $this->db->result();
+        $data = new stdClass();
+        $data->c = $comments;
+        $data->pid = $post_id;
+        return $data;
     }
